move(embedded.modes): into /feature
