+++ server/src/feature/embeddedLanguages/languageModes.ts
@@ -14,13 +14,13 @@ import {
 } from "vscode-html-languageservice";
 import { TextDocumentPositionParams } from "vscode-languageserver";
 import { DefinitionResult } from "../definition/getDefinition";
-import { getAttributeInterpolationMode } from "../../modes/getAttributeInterpolationMode";
-import { getAttributeMode } from "../../modes/getAttributeMode";
-import { getAureliaHtmlMode } from "../../modes/getAureliaHtmlMode";
-import { getCustomElementMode } from "../../modes/getCustomElementMode";
-import { getRepeatForMode } from "../../modes/getRepeatForMode";
-import { getTextInterpolationMode } from "../../modes/getTextInterpolationMode";
-import { getValueConverterMode } from "../../modes/getValueConverterMode";
+import { getAttributeInterpolationMode } from "./modes/getAttributeInterpolationMode";
+import { getAttributeMode } from "./modes/getAttributeMode";
+import { getAureliaHtmlMode } from "./modes/getAureliaHtmlMode";
+import { getCustomElementMode } from "./modes/getCustomElementMode";
+import { getRepeatForMode } from "./modes/getRepeatForMode";
+import { getTextInterpolationMode } from "./modes/getTextInterpolationMode";
+import { getValueConverterMode } from "./modes/getValueConverterMode";
 import { aureliaProgram } from "../../viewModel/AureliaProgram";
 import { AureliaCompletionItem } from "../../virtual/virtualCompletion/virtualCompletion";
 import {
