+++ server/src/feature/embeddedLanguages/modes/getAttributeMode.ts
@@ -1,28 +1,20 @@
-import {
-  ViewRegionInfo,
-  ViewRegionType,
-} from "../feature/embeddedLanguages/embeddedSupport";
+import { ViewRegionInfo, ViewRegionType } from "../embeddedSupport";
 import { TextDocumentPositionParams } from "vscode-languageserver";
-import { HTMLDocumentRegions } from "../feature/embeddedLanguages/embeddedSupport";
-import { LanguageModelCache } from "../feature/embeddedLanguages/languageModelCache";
-import {
-  LanguageMode,
-  Position,
-  TextDocument,
-} from "../feature/embeddedLanguages/languageModes";
-import { getAureliaVirtualCompletions } from "../virtual/virtualCompletion/virtualCompletion";
-import {
-  getAccessScopeDefinition,
-  getAccessScopeViewModelDefinition,
-} from "../feature/definition/accessScopeDefinition";
-import { DefinitionResult } from "../feature/definition/getDefinition";
+import { HTMLDocumentRegions } from "../embeddedSupport";
+import { LanguageModelCache } from "../languageModelCache";
+import { LanguageMode, Position, TextDocument } from "../languageModes";
+import { getAureliaVirtualCompletions } from "../../../virtual/virtualCompletion/virtualCompletion";
+import { DefinitionResult } from "../../definition/getDefinition";
+import { getVirtualDefinition } from "../../../virtual/virtualDefinition/virtualDefinition";
+import { aureliaProgram } from "../../../viewModel/AureliaProgram";
+import { getAccessScopeDefinition } from "../../definition/accessScopeDefinition";
 
-export function getRepeatForMode(
+export function getAttributeMode(
   documentRegions: LanguageModelCache<Promise<HTMLDocumentRegions>>
 ): LanguageMode {
   return {
     getId() {
-      return ViewRegionType.RepeatFor;
+      return ViewRegionType.Attribute;
     },
     async doComplete(
       document: TextDocument,
@@ -41,14 +33,14 @@ export function getRepeatForMode(
     async doDefinition(
       document: TextDocument,
       position: Position,
-      goToSourceWord: string,
-      attributeRegion: ViewRegionInfo
+      goToSourceWord: string
     ): Promise<DefinitionResult | undefined> {
       const regions = (await documentRegions.get(document)).getRegions();
-      return getAccessScopeViewModelDefinition(
+      return getAccessScopeDefinition(
         document,
         position,
-        goToSourceWord
+        goToSourceWord,
+        regions
       );
     },
     onDocumentRemoved(_document: TextDocument) {},
