+++ server/src/feature/embeddedLanguages/modes/getAureliaHtmlMode.ts
@@ -1,14 +1,10 @@
-import { ViewRegionType } from "../feature/embeddedLanguages/embeddedSupport";
+import { ViewRegionType } from "../embeddedSupport";
 import { TextDocumentPositionParams } from "vscode-languageserver";
-import { HTMLDocumentRegions } from "../feature/embeddedLanguages/embeddedSupport";
-import { LanguageModelCache } from "../feature/embeddedLanguages/languageModelCache";
-import {
-  LanguageMode,
-  Position,
-  TextDocument,
-} from "../feature/embeddedLanguages/languageModes";
-import { getAureliaVirtualCompletions } from "../virtual/virtualCompletion/virtualCompletion";
-import { aureliaProgram } from "../viewModel/AureliaProgram";
+import { HTMLDocumentRegions } from "../embeddedSupport";
+import { LanguageModelCache } from "../languageModelCache";
+import { LanguageMode, Position, TextDocument } from "../languageModes";
+import { getAureliaVirtualCompletions } from "../../../virtual/virtualCompletion/virtualCompletion";
+import { aureliaProgram } from "../../../viewModel/AureliaProgram";
 
 export function getAureliaHtmlMode(
   documentRegions: LanguageModelCache<Promise<HTMLDocumentRegions>>
